chore(optionSchemas): move up
