+++ packages/functions/src/alphabetical.ts
@@ -2,7 +2,7 @@ import { isObject } from 'lodash';
 import { createRulesetFunction } from '@stoplight/spectral-core';
 import { printValue } from '@stoplight/spectral-runtime';
 
-import { optionSchemas } from './schema/optionSchemas';
+import { optionSchemas } from './optionSchemas';
 
 export type Options = {
   /** if sorting array of objects, which key to use for comparison */
