+++ packages/functions/src/xor.ts
@@ -1,7 +1,7 @@
 import { createRulesetFunction, IFunctionResult } from '@stoplight/spectral-core';
 import { printValue } from '@stoplight/spectral-runtime';
 
-import { optionSchemas } from './schema/optionSchemas';
+import { optionSchemas } from './optionSchemas';
 
 export type Options = {
   /** test to verify if one (but not all) of the provided keys are present in object */
